+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New function `uef_maps:is_key_nested/2`
 - New function `uef_maps:new_nested/1`
 - New function `uef_maps:new_nested/2`
+- New function `uef_maps:put_nested/3`
 
 ## [1.6.1] - 2019-05-12